@@ -4,7 +4,7 @@ plugins {
 apply plugin: 'com.google.gms.google-services'
 
 android {
-    compileSdk 29
+    compileSdk 32
 
     defaultConfig {
         applicationId "com.example.qracutie"
